SDIT-1658 New version of alerts by prisoner POST and get by migrationId
This adds two new endpoints:
* Create alerts by prisoner number, the old version will be deleted when all environments use the new version
* Get a summary of prisoners migrated

This added offenderNo to the alert mapping table as a convenience column and also adds a new table that is just a list of prisoners migrated.

This all differs from usual pattern because we migrate a batch of alerts by prisoner, so we need to know both how many prisoners have been migrated and how many alerts have been migrated.