SDIT-2090 Contact phones can now have a DPS type
DPS Phone Ids are unique per type, ADDRESS or PERSON - unlike NOMIS that a globally unique.

This slightly breaks the pattern that all mapping entities have a primary key of DPS key that is a String, so for Phones switch to Nomis Id - which is a Long - to primary key.

Hence, a few changes t0 break out of the generic pattern.