SDIT-2868 (Don't) Perform batch updates during migrations (#613)
Some migrated prisoners have a lot of mapping data and so a lot of SQL inserts occurred inserting them individually. This change was supposed to batch up the entities and perform batch inserts - unfortunately batching is not supported by Spring R2DBC.

If we want batch inserts then we'll have to roll our own solution, hence it makes sense to keep this change and group the entities that could be batched. As a workaround for now the timeout in the migration service has been increased to 60s.

This change will provide a small increase in performance because we're no longer trying to upsert every single address, we're de-duplicating them first.